requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.6
django-cors-headers==2.2.0
django-rest-swagger==2.1.2
django-summernote==0.8.8.6
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==1.2
djoser==1.1.5
ecdsa==0.13.2
future==0.17.1
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==6.0.0
psycopg2==2.7.5
pyasn1==0.4.5
PyJWT==1.6.4
python-jose==3.0.1
pytz==2018.5
requests==2.19.1
rsa==4.0
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.23
